Move send_and_translate_response to lib.rs
